Rename utils.js to config.js
